+++ examples/widget_session_create.php
@@ -16,7 +16,18 @@ use Twizo\Examples\Util\EntityFormatter;
 
 $twizo = Twizo\Api\Twizo::getInstance(SECRET, API_HOST);
 
-$widget = $twizo->createWidgetSession(readline('Number: ' ));
+$allowedTypes = array_map('trim', explode(',', readline('AllowedType(s) (Delimiter: ","): ')));
+$recipient = null;
+$backupCodeIdentifier = null;
+if (in_array(WidgetSession::TYPE_SMS, $allowedTypes) || in_array(WidgetSession::TYPE_CALL, $allowedTypes)) {
+    $recipient = readline('Number: ');
+}
+
+if (in_array(WidgetSession::TYPE_BACKUP_CODE, $allowedTypes)) {
+    $backupCodeIdentifier = readline('BackupCodeIdentifier: ');
+}
+
+$widget = $twizo->createWidgetSession($allowedTypes, $recipient, $backupCodeIdentifier);
 $widget->create();